Update Numbers field to allow either Digits or Number validation and sanitize as float instead of int
git-svn-id: https://plugins.svn.wordpress.org/visual-form-builder/trunk@709970 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ visual-form-builder.php
@@ -1989,7 +1989,7 @@ class Visual_Form_Builder{
 				break;
 				
 				case 'number' :
-					return intval( $data );
+					return floatval( $data );
 				break;
 				
 				case 'address' :
